+++ index.html
@@ -11,8 +11,7 @@
 
 <body>
   <main>
-    <script type="text/javascript" src="./js/main.js" async></script>
-    <script type="text/javascript" src="./js/functions.js" async></script>
+    <script src="./js/main.js" type="module" async></script>
     <!-- Фильтрация изображений от других пользователей -->
     <section class="img-filters  img-filters--inactive  container">
       <h2 class="img-filters__title  visually-hidden">Фильтр фотографий</h2>
